refactor: split commands/remote_copy.rs into legacy + serve submodules
remote_copy.rs (928) → remote_copy.rs (316) + remote_copy/
  legacy.rs (328) — handle_remote_upload, handle_remote_download
  serve.rs  (305) — handle_serve_upload, serve_upload_dir,
                    handle_serve_download

Top-level keeps the shared helpers + handle_remote_copy dispatcher;
pub(super) widens just enough for cross-submodule calls. Public API
(handle_remote_copy, is_plain_mode) resolves unchanged via existing
path. Last of the "still too large" files is done.
